Rename shared_helpers for Sentry for consistency
