+++ spec/lib/saml/settings_service_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 require 'saml/settings_service'
-require 'lib/sentry_logging_spec_helper'
+require 'lib/shared_examples_for_sentry_logging'
 
 RSpec.describe SAML::SettingsService do
   before do
