+++ spec/lib/shared_examples_for_sentry_logging.rb
